Added basic input system
Any entity/node can receive signals from input but only after it
becomes registered in controller. To do that, it needs a component
called Input Register which can either be active (registers on
ready) or inactive (waits for manual register). Once register exits
the tree, it calls controller to remove it from registered entities
list.

Scenes:
- Input Register

Add it as child to any entity to give it right to process input.

- Start input

Simple input that enables skipping intro or starting the game.

Globals:
- Input Controller

Class which stores registered entity references and processes
input gathering for each of them.