Added string wrapper
String wrapper is used in Dialogue system to enable calculating
size of characters to show during character by character reveal.
This couldn't be calculated beforehand as currently (Godot 4.3)
engine misses functionality to scroll text pixel by pixel and to
do so manually, I don't have access to visible line or character
values.

Wrapper takes any basic string and reconstructs it in a way that
we are wrapping words and dividing complex words (words that have
bigger length than max chars per line) in a way original does.

Excluding complex white strings cases (which are also included
and tested), basically we try to fit text in number of lines and
if we can't we move the word to next line. Each sentence is not
combined with another and starts from the new line.

Original Pokemon dialogue box has set max text length per line
as 16 characters and I keep i this way.