Added Limiter
Limiter is a utility class that wraps counter, timer and signal.
It's purpose is to manage various object's lifetime. Limiter's
constructor is deducing from variant which type we are using.
It can also take existing Counter/Timer as argument and use those
as limit handlers.

Limiter was properly tested and documented.